test: add manual SPI full-duplex loopback checks
